+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 更改日志 | Change Log
 
 - 初始化项目 | Initialize project
+- API 测试支持使用 JSON Schema 校验响应体 | API test supports using JSON Schema to verify response body